each thread operates on own set of rows

+++ kaczmarz/src/solvers/asynchronous.hpp
@@ -1,7 +1,7 @@
 #include <omp.h>
 #include <unistd.h>
 
-#include "../kaczmarz_serial/kaczmarz_common.hpp"
+#include "common.hpp"
 #include "linear_systems/sparse.hpp"
 #include "linear_systems/types.hpp"
 
@@ -9,4 +9,4 @@ KaczmarzSolverStatus sparse_kaczmarz_parallel(const SparseLinearSystem &lse,
                                               Vector &x,
                                               const unsigned max_iterations,
                                               const double precision,
-                                              const double num_threads);
+                                              const unsigned num_threads);
